requirements.txt
rasterio
geopandas
shapely
scipy
Pillow
